Brute force split quote implementation
